Kernel with a different config based on a version 3.13 kernel
This build is based on a config without modules enabled and a more optimal config file for kernel version 3.13 for the chromebook snow from virtual open systems (vos) and is updated to kernel version 5.4 using the make oldconfig target when building the kernel. The full config file is included in the tar.gz archive. Testing has indicated that with the kernel 3.13 and the base 3.13 config file, a chromebook snow with kvm enabled will have access to the full 2 GB of memory, while the default config of kernel version 5.4 with kvm enabled only has access to 1 GB of the 2 GB available on the device. This upload is to test this new configuration to try to enable the full 2 GB with kvm enabled and a version 5.4 kernel.